updated codes for recurrent payment

+++ src/main/java/com/wayapaychat/paymentgateway/pojo/unifiedpayment/UnifiedPaymentRecurrentPaymentRequest.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Data
 @Component
 @NoArgsConstructor
@@ -14,8 +16,8 @@ public class UnifiedPaymentRecurrentPaymentRequest {
     @Value("${waya.unified-payment.secret}")
     private String SecretKey;
     private String Scheme;
-    private Float Amount;
-    private Float Fee;
+    private BigDecimal Amount;
+    private Integer Fee;
     private String SessionId;
     private String Currency = "566";
     private String CustomerEmail;
